==> Step 1 - PrepareDataset.ipynb <==
from time import sleep
import pandas as pd

df = pd.read_csv("./data/dataset.csv")
df.dropna(inplace=True)
df = df.apply(lambda x: x.str.lower())
df["Title_Clean"] = df.page_title.apply(lambda x: x.split(" \\ ")[-1].strip())
df.head(n=15)

mlm_df = df[["Title_Clean", "content"]]
mlm_df.to_csv("./data/mlm_dataset.csv", index=False)
