==> Step 2 - ExtractVocabulary.ipynb <==
from tokenizers import BertWordPieceTokenizer
from pathlib import Path
import pandas as pd


dataset = pd.read_csv("./data/mlm_dataset.csv")


dataset.content.to_list()

dataset["content"].to_csv("./data/mlm_dataset.txt", index=False)

VOCAB_SIZE = 30522

# Read input files from local input path 
paths = [str(x) for x in Path("./").glob('*.txt')]

# Train custom BertWordPiece tokenizer
tokenizer = BertWordPieceTokenizer()
tokenizer.train(files=paths, vocab_size=VOCAB_SIZE)

# Save trained custom tokenizer to local output path
tokenizer.save_model("./data/vocab")

tokenizer = BertWordPieceTokenizer('./data/vocab/vocab.txt')



# Evaluate custom tokenizer 
print('Evaluating custom tokenizer')
test_sentence = 'amazon simple storage service (amazon s3) is an object storage service that offers industry-leading scalability, data availability, security, and performance.'
print(f'Test sentence: {test_sentence}')
tokens = tokenizer.encode(test_sentence).tokens
print(f'Encoded sentence: {tokens}')
token_id = tokenizer.token_to_id('s3')
print(f'Token ID for token (s3) = {token_id}')
vocab_size = tokenizer.get_vocab_size()
print(f'Vocabulary size = {vocab_size}')

tokenizer.encode(test_sentence).tokens

tokenizer.save_model("./data/vocab")
