==> Step 4 - FineTune.ipynb <==
from transformers import DataCollatorForLanguageModeling
from transformers import TrainingArguments
from transformers import BertTokenizerFast
from transformers import BertForMaskedLM
from transformers import BertConfig
from transformers import pipeline 
from datasets import load_dataset
from transformers import Trainer
from datasets import DatasetDict
import pandas as pd
import transformers
import datasets
import logging
import torch
import math
import sys


# Setup logging
logger = logging.getLogger(__name__)
logging.basicConfig(level=logging.getLevelName('INFO'), 
                    handlers=[logging.StreamHandler(sys.stdout)], 
                    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')

# Log versions of dependencies
logger.info(f'[Using Transformers: {transformers.__version__}]')
logger.info(f'[Using Datasets: {datasets.__version__}]')
logger.info(f'[Using Torch: {torch.__version__}]')
logger.info(f'[Using Pandas: {pd.__version__}]')


MAX_LENGTH = 512
# Re-create BERT WordPiece tokenizer 
logger.info(f'Re-creating BERT tokenizer using custom vocabulary from [./data/vocab/]')
# config = BertConfig()
tokenizer = BertTokenizerFast.from_pretrained("./data/vocab",padding="max_length", truncation="max_length")
tokenizer.model_max_length = MAX_LENGTH
tokenizer.init_kwargs['model_max_length'] = MAX_LENGTH
tokenizer.save_pretrained("./BTF")
logger.info(f'Tokenizer: {tokenizer}')


MAX_LENGTH = 512
CHUNK_SIZE = 128
TRAIN_EPOCHS = 5
BATCH_SIZE = 128
SAVE_STEPS = 10000
SAVE_TOTAL_LIMIT = 2

LOCAL_DATA_DIR = './'
LOCAL_MODEL_DIR = './finetuned'




# # Re-create original BERT WordPiece tokenizer 
# logger.info(f'Re-creating original BERT tokenizer')
# tokenizer = BertTokenizerFast.from_pretrained('bert-base-uncased')
# logger.info(f'Tokenizer: {tokenizer}')

# Re-create BERT WordPiece tokenizer 
logger.info(f'Re-creating BERT tokenizer using custom vocabulary from [./data/vocab/]')
config = BertConfig()
tokenizer = BertTokenizerFast.from_pretrained("./data/vocab", config=config, padding="max_length", truncation="max_length")
tokenizer.model_max_length = MAX_LENGTH
tokenizer.init_kwargs['model_max_length'] = MAX_LENGTH
tokenizer.save_model("./BTF")
logger.info(f'Tokenizer: {tokenizer}')

# Read dataset 
chunked_datasets = datasets.load_from_disk(LOCAL_DATA_DIR)
logger.info(f'Chunked datasets: {chunked_datasets}')

# Create data collator
data_collator = DataCollatorForLanguageModeling(tokenizer=tokenizer, 
                                                mlm=True, 
                                                mlm_probability=0.15)
    
# Load MLM
logger.info('Loading BertForMaskedLM model')
mlm = BertForMaskedLM(config=config)

# Train MLM
logger.info('Training MLM')
training_args = TrainingArguments(output_dir='/tmp/checkpoints', 
                                    overwrite_output_dir=True, 
                                    optim='adamw_torch',
                                    num_train_epochs=TRAIN_EPOCHS,
                                    per_device_train_batch_size=BATCH_SIZE,
                                    evaluation_strategy='epoch',
                                    save_steps=SAVE_STEPS, 
                                    save_total_limit=SAVE_TOTAL_LIMIT)
trainer = Trainer(model=mlm, 
                    args=training_args, 
                    data_collator=data_collator,
                    train_dataset=chunked_datasets['train'],
                    eval_dataset=chunked_datasets['validation'])

# Evaluate trained model for perplexity
eval_results = trainer.evaluate()
logger.info(f"Perplexity before training: {math.exp(eval_results['eval_loss']):.2f}")

trainer.train()

eval_results = trainer.evaluate()
logger.info(f"Perplexity after training: {math.exp(eval_results['eval_loss']):.2f}")

# Save trained model to local model directory
logger.info(f'Saving trained MLM to [{LOCAL_MODEL_DIR}/]')
trainer.save_model(LOCAL_MODEL_DIR)

eval_results
